+++ apps/web/package.json
@@ -23,7 +23,9 @@
     "@gekichumai/oneshot": "*",
     "@mui/material": "^5.14.13",
     "@sentry/react": "^7.98.0",
-    "@supabase/supabase-js": "^2.39.3",
+    "@supabase/auth-ui-react": "^0.4.7",
+    "@supabase/auth-ui-shared": "^0.1.8",
+    "@supabase/supabase-js": "^2.39.5",
     "@tanstack/react-table": "^8.10.7",
     "@unocss/preset-mini": "^0.58.4",
     "@unocss/preset-wind": "^0.58.4",
