+++ apps/web/src/components/global/DXRank.tsx
@@ -1,6 +1,6 @@
 import clsx from "clsx";
 import { FC } from "react";
-import { useAppContextSlugVersion } from "../models/context/useAppContext";
+import { useAppContextSlugVersion } from "../../models/context/useAppContext";
 
 export const DXRank: FC<{ rank?: string | null; className?: string }> = ({
   rank,
