+++ apps/web/src/components/rating/io/ImportFromAquaSQLiteButton.tsx
@@ -32,8 +32,8 @@ import {
   readAquaPlayLogs,
   readAquaUsers,
 } from "../../../utils/aquaDB";
+import { SheetListItemContent } from "../../sheet/SheetListItem";
 import { PlayEntry } from "../RatingCalculatorAddEntryForm";
-import { SheetListItemContent } from "../../SheetListItem";
 
 import IconMdiDatabase from "~icons/mdi/database";
 import { FadedImage } from "../../global/FadedImage";
