+++ apps/web/src/components/sheet/SheetListContainer.tsx
@@ -1,7 +1,7 @@
 import clsx from "clsx";
 import { useCallback } from "react";
 import { ItemContent, Virtuoso } from "react-virtuoso";
-import { FlattenedSheet } from "../songs";
+import { FlattenedSheet } from "../../songs";
 import { SheetListItem } from "./SheetListItem";
 
 export const SheetListContainer = ({
