+++ apps/web/src/components/sheet/SheetListItem.tsx
@@ -9,15 +9,15 @@ import { FC, HTMLAttributes, ImgHTMLAttributes, memo, useState } from "react";
 import toast from "react-hot-toast";
 import { match } from "ts-pattern";
 import MdiComment from "~icons/mdi/comment";
-import { DIFFICULTIES } from "../models/difficulties";
-import { FlattenedSheet } from "../songs";
-import { useIsLargeDevice } from "../utils/breakpoints";
-import { FadedImage } from "./global/FadedImage";
-import { ResponsiveDialog } from "./global/ResponsiveDialog";
+import { DIFFICULTIES } from "../../models/difficulties";
+import { FlattenedSheet } from "../../songs";
+import { useIsLargeDevice } from "../../utils/breakpoints";
+import { FadedImage } from "../global/FadedImage";
+import { ResponsiveDialog } from "../global/ResponsiveDialog";
 import {
   SheetDialogContent,
   SheetDialogContentProps,
-} from "./sheet/SheetDialogContent";
+} from "./SheetDialogContent";
 
 export const SheetListItem: FC<
   {
